Move the ads library to 7.0.317 so one build serves 4024 and 4026
On a pc running TwinCAT 4026 the tool could not connect at all, failing with

  AdsException: Cannot register Port '0'

against any target, including the pc itself. The 5.x library reaches the local
router only over tcp loopback on 127.0.0.1:48898, and 4026 changed that layer -
it can use a unix socket instead. The same executable worked against a 4024
machine, which is what showed the code was not at fault.

Of the available lines, 5.x stops at 4024, 6.x requires 4024.10 or later, and
7.x supports TwinCAT 2.11 and up. Only 7.x covers both with a single binary.
System.Reactive moves to 6.1.0, which 7.x requires.

Three compilation points: DT.Date became DT.Value, DATE.Date changed type, and
TcVersion.Build is now a short.

The change the compiler does not report is that the PlcOpen types hand back a
DateTime instead of a DateTimeOffset. The offset was an artefact of the wrapper
- a plc DT carries no time zone - so a backup now records the timestamp as the
plc holds it. Verified by running the library: DT -> Normalize -> json ->
TryCoerce -> DT gives back the same ticks, and the DT(DateTimeOffset) and
TIME(TimeSpan) constructors still exist, so backups written with 5.x remain
restorable.

One test was updated deliberately: it asserted that Normalize of a DT returned a
DateTimeOffset, which is exactly the type that changed. The ones that check the
instant and the round trip are untouched and pass.

+++ TwinCatAdsTool.Logic/Services/DeviceFinder.cs
@@ -131,7 +131,7 @@ namespace TwinCatAdsTool.Logic.Services
 
                     device.TcVersion.Version = tail[i + 4];
                     device.TcVersion.Revision = tail[i + 5];
-                    device.TcVersion.Build = tail[i + 6] + tail[i + 7] * 256;
+                    device.TcVersion.Build = (short) (tail[i + 6] + tail[i + 7] * 256);
                     break;
                 }
             }
